Deduplicate ExperimentConfig/read_config; move initial_condition to common
`ExperimentConfig` and the Fortran-namelist reader were duplicated between
`standalone_driver` and `testing`, each marked "duplicate ... to avoid circular
imports". The duplication existed because `testing` depended on
`standalone_driver` (for `DriverConfig`/`InitialConditionConfig`), which is the
wrong direction: the standalone driver should sit at the top of the dependency
graph, with nothing depending on it.

Changes:
- New `icon4py.model.experiment_config` package holds the single `ExperimentConfig`,
  `DriverConfig`, and the shared `read_experiment_config` reader. It sits above the
  atmosphere packages but below both `testing` and `standalone_driver`, so neither
  depends on the other. The `testing -> standalone_driver` edge is removed.
- `standalone_driver.config.read_config` and
  `testing.datatest_utils.create_experiment_configuration` become thin wrappers
  over the shared reader.
- Move the `initial_condition` subpackage to `common/initial_condition/`, mirroring
  `common/topography/`, so the IC configs stay next to their compute functions.
  This required moving `StaticFieldFactories` (a NamedTuple of common factories)
  to `common/states/static_fields.py`.
- Drop the now-unused `serialbox4py` dependency from `standalone_driver` (its only
  user moved to `common`, which already depends on it).
- Remove the now-resolved `# type: ignore[arg-type]` / `TODO(1320)` workarounds in
  the standalone driver initial-condition tests.
