+++ model/common/src/icon4py/model/common/initial_condition/__init__.py
@@ -6,7 +6,7 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
-from icon4py.model.standalone_driver.initial_condition.config import InitialConditionConfig, create
+from icon4py.model.common.initial_condition.config import InitialConditionConfig, create
 
 
 __all__ = [
