+++ model/standalone_driver/src/icon4py/model/standalone_driver/standalone_driver.py
@@ -22,7 +22,13 @@ import icon4py.model.common.utils as common_utils
 from icon4py.model.atmosphere.advection import advection_states
 from icon4py.model.atmosphere.diffusion import diffusion_states
 from icon4py.model.atmosphere.dycore import dycore_states
-from icon4py.model.common import dimension as dims, model_backends, topography, type_alias as ta
+from icon4py.model.common import (
+    dimension as dims,
+    initial_condition,
+    model_backends,
+    topography,
+    type_alias as ta,
+)
 from icon4py.model.common.decomposition import definitions as decomposition_defs
 from icon4py.model.common.grid import (
     geometry_attributes as geom_attr,
@@ -44,7 +50,6 @@ from icon4py.model.standalone_driver import (
     driver_io,
     driver_states,
     driver_utils,
-    initial_condition,
 )
