+++ model/testing/src/icon4py/model/testing/datatest_utils.py
@@ -8,29 +8,14 @@
 
 from __future__ import annotations
 
-import dataclasses
-import json
 import logging
 import pathlib
 import urllib.parse
 
 import gt4py.next.typing as gtx_typing
 
-from icon4py.model.atmosphere.advection import advection as tracer_advection
-from icon4py.model.atmosphere.diffusion import diffusion
-from icon4py.model.atmosphere.dycore import solve_nonhydro as solve_nh
-from icon4py.model.atmosphere.subgrid_scale_physics.microphysics import (
-    single_moment_six_class_gscp_graupel as graupel,
-)
-from icon4py.model.common import topography
 from icon4py.model.common.decomposition import definitions as decomposition
-from icon4py.model.common.grid import vertical as v_grid
-from icon4py.model.common.interpolation import interpolation_factory
-from icon4py.model.common.metrics import metrics_factory
-from icon4py.model.common.states import tracer_state
-from icon4py.model.common.utils import fortran_config
-from icon4py.model.standalone_driver import config as driver_config, initial_condition
-from icon4py.model.standalone_driver.initial_condition import from_file as from_file_ic
+from icon4py.model.experiment_config.reader import read_experiment_config
 from icon4py.model.testing import data_handling, definitions, serialbox
 
 
@@ -151,97 +136,5 @@ def create_experiment_configuration(
     experiment_description: definitions.ExperimentDescription,
     processor_props: decomposition.ProcessProperties,
 ) -> definitions.ExperimentConfig:
-    # NOTE: This has a duplicate in standalone_driver/config.py to avoid circular imports.
-
-    experiment_path = get_path_for_experiment(
-        experiment_description,
-        processor_props,
-    )
-
-    with (experiment_path / fortran_config.ATM_DICT_FNAME).open() as f:
-        atm_dict = json.load(f)
-    with (experiment_path / fortran_config.MASTER_DICT_FNAME).open() as f:
-        master_dict = json.load(f)
-    with (experiment_path / fortran_config.INPUT_DICT_FNAME).open() as f:
-        input_dict = json.load(f)
-
-    metrics_config = metrics_factory.MetricsConfig.from_fortran_dict(atm_dict)
-
-    interpolation_config = interpolation_factory.InterpolationConfig.from_fortran_dict(atm_dict)
-
-    vertical_grid_config = v_grid.VerticalGridConfig.from_fortran_dict(atm_dict)
-
-    topography_config = topography.TopographyConfig.from_fortran_dict(
-        atm_dict=atm_dict, input_dict=input_dict, data_path=experiment_path
-    )
-
-    nonhydro_config = solve_nh.NonHydrostaticConfig.from_fortran_dict(
-        atm_dict,
-        max_nudging_coefficient=interpolation_config.max_nudging_coefficient,
-    )
-
-    diffusion_config = diffusion.DiffusionConfig.from_fortran_dict(
-        atm_dict,
-        max_nudging_coefficient=interpolation_config.max_nudging_coefficient,
-    )
-
-    do_tracer_advection = experiment_description not in (
-        definitions.Experiments.MCH_CH_R04B09,
-        definitions.Experiments.EXCLAIM_APE,
-    )
-    # The experiments above were run in fortran with a tracer advection scheme
-    # that has not been ported to ICON4Py and can not be used for testing.
-    # TODO (jcanton): this isn't the right place to keep a special case
-    # handling. Either fix these experiments or move the special case handling.
-    tracer_advection_config = (
-        tracer_advection.AdvectionConfig.from_fortran_dict(atm_dict)
-        if do_tracer_advection
-        else None
-    )
-    ntracer = (
-        fortran_config.list_to_value(atm_dict["run_nml"]["ntracer"]) if do_tracer_advection else 0
-    )
-    tracer_config = tracer_state.TracerConfig.from_ntracer(ntracer)
-
-    do_physics = "nwp_phy_nml" in atm_dict and "nwp_tuning_nml" in atm_dict
-    # If these two namelists are missing it means that the experiment was run
-    # without microphysics and we have to skip parsing the graupel config which
-    # relies on some of these parameters.
-    graupel_config = (
-        graupel.SingleMomentSixClassIconGraupelConfig.from_fortran_dict(atm_dict)
-        if do_physics
-        else None
-    )
-
-    initial_condition_config = initial_condition.InitialConditionConfig.from_fortran_dict(
-        atm_dict=atm_dict, input_dict=input_dict, data_path=experiment_path
-    )
-
-    if not do_tracer_advection and isinstance(
-        initial_condition_config.config, from_file_ic.FromFileConfig
-    ):
-        initial_condition_config = dataclasses.replace(
-            initial_condition_config,
-            config=dataclasses.replace(initial_condition_config.config, ntracer=0),
-        )
-
-    driver_cfg = driver_config.DriverConfig.from_fortran_dict(
-        atm_dict=atm_dict,
-        master_dict=master_dict,
-        profiling_stats=None,
-        enable_statistics_output=False,
-    )
-
-    return definitions.ExperimentConfig(
-        metrics=metrics_config,
-        interpolation=interpolation_config,
-        vertical_grid=vertical_grid_config,
-        topography=topography_config,
-        nonhydrostatic=nonhydro_config,
-        diffusion=diffusion_config,
-        tracer_config=tracer_config,
-        tracer_advection=tracer_advection_config,
-        graupel=graupel_config,
-        initial_condition=initial_condition_config,
-        driver=driver_cfg,
-    )
+    experiment_path = get_path_for_experiment(experiment_description, processor_props)
+    return read_experiment_config(experiment_path)
