+++ model/testing/src/icon4py/model/testing/definitions.py
@@ -11,24 +11,11 @@ from __future__ import annotations
 import copy
 import dataclasses
 import pathlib
-from typing import TYPE_CHECKING, Any, Final
+from typing import Final
 
-from icon4py.model.atmosphere.advection import advection as tracer_advection
-from icon4py.model.atmosphere.subgrid_scale_physics.microphysics import (
-    single_moment_six_class_gscp_graupel as graupel,
-)
-from icon4py.model.common import topography
-from icon4py.model.common.grid import icon as icon_grid, vertical as v_grid
-from icon4py.model.common.interpolation import interpolation_factory
-from icon4py.model.common.metrics import metrics_factory
-from icon4py.model.common.states import tracer_state
-from icon4py.model.standalone_driver import config as driver_config, initial_condition
+from icon4py.model.common.grid import icon as icon_grid
+from icon4py.model.experiment_config.config import ExperimentConfig
 from icon4py.model.testing import config
-
-
-if TYPE_CHECKING:
-    from icon4py.model.atmosphere.diffusion import diffusion
-    from icon4py.model.atmosphere.dycore import solve_nonhydro as solve_nh
 
 
 SERIALIZED_DATA_DIR: Final = "ser_icondata"
@@ -170,32 +157,6 @@ class ExperimentDescription:
     long_name: str
     grid: GridDescription
     version: int = 5
-
-
-@dataclasses.dataclass
-class ExperimentConfig:
-    # NOTE: This has a duplicate in standalone_driver/config.py to avoid circular imports.
-    metrics: metrics_factory.MetricsConfig
-    interpolation: interpolation_factory.InterpolationConfig
-    vertical_grid: v_grid.VerticalGridConfig
-    topography: topography.TopographyConfig
-    initial_condition: initial_condition.InitialConditionConfig
-    driver: driver_config.DriverConfig
-    nonhydrostatic: solve_nh.NonHydrostaticConfig | None = None
-    diffusion: diffusion.DiffusionConfig | None = None
-    tracer_config: tracer_state.TracerConfig | None = None
-    tracer_advection: tracer_advection.AdvectionConfig | None = None
-    graupel: graupel.SingleMomentSixClassIconGraupelConfig | None = None
-
-    def with_overrides(self, **overrides: Any) -> ExperimentConfig:
-        replacements: dict[str, Any] = {}
-        for key, value in overrides.items():
-            current = getattr(self, key)
-            if isinstance(value, dict):
-                replacements[key] = dataclasses.replace(current, **value)
-            else:
-                replacements[key] = value
-        return dataclasses.replace(self, **replacements)
 
 
 @dataclasses.dataclass
